automation/social_bots.py
"""
Social Media / Messaging Bot Automation Module

This module provides functionality for automating social media interactions
across multiple platforms including Telegram, Discord, and Twitter.

Features:
- Telegram / Discord / Twitter/X bots
- Auto post, message, fetch, or monitor
- OAuth / token management
- Configurable triggers and message templates
- Async, retry, logging, storage support

Example:
    config = {
        "social_bots": {
            "telegram": {
                "token": "YOUR_TELEGRAM_BOT_TOKEN",
                "chat_id": "YOUR_CHAT_ID"
            }
        }
    }
    social_bots = SocialBots(config)
    results = await social_bots.run()
"""

import asyncio
import logging
import json
import tweepy
from typing import Dict, List, Any, Optional
from datetime import datetime
import random
import time

# Telegram bot support
try:
    from telegram import Bot
    from telegram.error import TelegramError
except ImportError:
    Bot = None
    TelegramError = Exception

# Discord bot support
try:
    import discord
    from discord.ext import commands
except ImportError:
    discord = None
    commands = None

class SocialBots:
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.social_configs = config.get('social_bots', {})
        
        # Setup logging
        log_config = config.get('logging', {})
        logging.basicConfig(
            level=getattr(logging, log_config.get('level', 'INFO')),
            format='%(asctime)s - %(levelname)s - %(message)s',
            handlers=[
                logging.FileHandler(log_config.get('file', 'social_bots.log')),
                logging.StreamHandler()
            ]
        )
        self.logger = logging.getLogger(__name__)
        
        # Initialize clients
        self.telegram_bot = None
        self.discord_bot = None
        self.twitter_api = None
        
        self.init_bots()
    
    def init_bots(self):
        """Initialize social media bots"""
        # Initialize Telegram bot
        telegram_config = self.social_configs.get('telegram', {})
        if telegram_config and Bot:
            try:
                self.telegram_bot = Bot(token=telegram_config.get('token', ''))
                self.logger.info("Telegram bot initialized")
            except Exception as e:
                self.logger.error(f"Failed to initialize Telegram bot: {str(e)}")
        
        # Initialize Twitter API
        twitter_config = self.social_configs.get('twitter', {})
        if twitter_config:
            try:
                consumer_key = twitter_config.get('consumer_key', '')
                consumer_secret = twitter_config.get('consumer_secret', '')
                access_token = twitter_config.get('access_token', '')
                access_token_secret = twitter_config.get('access_token_secret', '')
                
                # Twitter API v2
                self.twitter_client = tweepy.Client(
                    consumer_key=consumer_key,
                    consumer_secret=consumer_secret,
                    access_token=access_token,
                    access_token_secret=access_token_secret
                )
                self.logger.info("Twitter API initialized")
            except Exception as e:
                self.logger.error(f"Failed to initialize Twitter API: {str(e)}")
    
    async def send_telegram_message(self, chat_id: str, message: str) -> Dict[str, Any]:
        """Send a message via Telegram bot"""
        if not self.telegram_bot:
            return {
                'platform': 'telegram',
                'status': 'error',
                'error': 'Telegram bot not initialized'
            }
        
        try:
            await self.telegram_bot.send_message(chat_id=chat_id, text=message)
            self.logger.info(f"Telegram message sent to {chat_id}")
            return {
                'platform': 'telegram',
                'status': 'success',
                'chat_id': chat_id
            }
        except TelegramError as e:
            self.logger.error(f"Failed to send Telegram message: {str(e)}")
            return {
                'platform': 'telegram',
                'status': 'error',
                'error': str(e),
                'chat_id': chat_id
            }
        except Exception as e:
            self.logger.error(f"Unexpected error sending Telegram message: {str(e)}")
            return {
                'platform': 'telegram',
                'status': 'error',
                'error': str(e),
                'chat_id': chat_id
            }
    
    def post_tweet(self, message: str) -> Dict[str, Any]:
        """Post a tweet"""
        if not self.twitter_client:
            return {
                'platform': 'twitter',
                'status': 'error',
                'error': 'Twitter API not initialized'
            }
        
        try:
            response = self.twitter_client.create_tweet(text=message)
            self.logger.info(f"Tweet posted with ID: {response.data['id']}")
            return {
                'platform': 'twitter',
                'status': 'success',
                'tweet_id': response.data['id']
            }
        except Exception as e:
            self.logger.error(f"Failed to post tweet: {str(e)}")
            return {
                'platform': 'twitter',
                'status': 'error',
                'error': str(e)
            }
    
    async def send_discord_message(self, channel_id: int, message: str) -> Dict[str, Any]:
        """Send a message via Discord bot"""
        # Note: This is a simplified implementation
        # A full Discord bot would require more setup
        return {
            'platform': 'discord',
            'status': 'error',
            'error': 'Discord bot implementation requires additional setup'
        }
    
    def load_message_templates(self, template_file: str) -> List[str]:
        """Load message templates from a file"""
        try:
            with open(template_file, 'r', encoding='utf-8') as f:
                templates = [line.strip() for line in f if line.strip()]
            return templates
        except Exception as e:
            self.logger.error(f"Failed to load message templates: {str(e)}")
            return []
    
    def select_random_message(self, templates: List[str], data: Dict[str, Any] = None) -> str:
        """Select a random message template and format it with data"""
        if not templates:
            return "Default message"
        
        message = random.choice(templates)
        
        # Simple placeholder replacement
        if data:
            for key, value in data.items():
                message = message.replace(f"{{{key}}}", str(value))
        
        return message
    
    async def run_telegram_tasks(self, tasks: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Run Telegram bot tasks"""
        results = []
        
        for task in tasks:
            task_type = task.get('type')
            
            if task_type == 'send_message':
                chat_id = task.get('chat_id')
                message = task.get('message', '')
                template_file = task.get('template_file')
                data = task.get('data', {})
                
                # Use template if provided
                if template_file:
                    templates = self.load_message_templates(template_file)
                    message = self.select_random_message(templates, data)
                
                result = await self.send_telegram_message(chat_id, message)
                results.append(result)
                
                # Add delay to respect rate limits
                delay = task.get('delay', 1)
                await asyncio.sleep(delay)
            
            else:
                results.append({
                    'platform': 'telegram',
                    'status': 'error',
                    'error': f'Unknown task type: {task_type}'
                })
        
        return results
    
    def run_twitter_tasks(self, tasks: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Run Twitter bot tasks"""
        results = []
        
        for task in tasks:
            task_type = task.get('type')
            
            if task_type == 'post_tweet':
                message = task.get('message', '')
                template_file = task.get('template_file')
                data = task.get('data', {})
                
                # Use template if provided
                if template_file:
                    templates = self.load_message_templates(template_file)
                    message = self.select_random_message(templates, data)
                
                result = self.post_tweet(message)
                results.append(result)
                
                # Add delay to respect rate limits
                delay = task.get('delay', 1)
                time.sleep(delay)
            
            else:
                results.append({
                    'platform': 'twitter',
                    'status': 'error',
                    'error': f'Unknown task type: {task_type}'
                })
        
        return results
    
    async def run_discord_tasks(self, tasks: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Run Discord bot tasks"""
        results = []
        
        for task in tasks:
            task_type = task.get('type')
            
            if task_type == 'send_message':
                channel_id = task.get('channel_id')
                message = task.get('message', '')
                template_file = task.get('template_file')
                data = task.get('data', {})
                
                # Use template if provided
                if template_file:
                    templates = self.load_message_templates(template_file)
                    message = self.select_random_message(templates, data)
                
                result = await self.send_discord_message(channel_id, message)
                results.append(result)
                
                # Add delay to respect rate limits
                delay = task.get('delay', 1)
                await asyncio.sleep(delay)
            
            else:
                results.append({
                    'platform': 'discord',
                    'status': 'error',
                    'error': f'Unknown task type: {task_type}'
                })
        
        return results
    
    async def run(self, dry_run: bool = False) -> List[Dict[str, Any]]:
        """Main execution method"""
        self.logger.info("Starting social bots automation")
        
        if dry_run:
            self.logger.info("DRY RUN: No actual social bot actions will be performed")
            return []
        
        all_results = []
        
        # Run Telegram tasks
        telegram_tasks = self.social_configs.get('telegram', {}).get('tasks', [])
        if telegram_tasks:
            telegram_results = await self.run_telegram_tasks(telegram_tasks)
            all_results.extend(telegram_results)
        
        # Run Twitter tasks
        twitter_tasks = self.social_configs.get('twitter', {}).get('tasks', [])
        if twitter_tasks:
            twitter_results = self.run_twitter_tasks(twitter_tasks)
            all_results.extend(twitter_results)
        
        # Run Discord tasks
        discord_tasks = self.social_configs.get('discord', {}).get('tasks', [])
        if discord_tasks:
            discord_results = await self.run_discord_tasks(discord_tasks)
            all_results.extend(discord_results)
        
        return all_results